% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_db.R
\docType{class}
\name{class_db}
\alias{class_db}
\title{Class describing object of type db}
\format{\code{\link{R6Class}} object.}
\usage{
class_db
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Class describing object of type db
}
\section{Methods}{


\describe{
  \item{\code{method()}}{description}
  }
}

\keyword{data}
