% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_db_basic.R
\docType{class}
\name{class_db_basic}
\alias{class_db_basic}
\title{Class describing object of type db}
\format{\code{\link{R6Class}} object.}
\usage{
class_db_basic
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Class describing object of type db
}
\section{Fields}{

\describe{
\item{\code{con}}{}
}}

\section{Methods}{


\describe{
  \item{\code{gallery(imgpathvector)}}{This method uploads images from vector of file paths \code{imgpathvector} to the server and creates a gallery of these images.}
  }
}

\keyword{data}
