% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_db_table_prototype.R
\docType{class}
\name{class_db_table_prototype}
\alias{class_db_table_prototype}
\title{Class providing object with methods for communication with lightning-viz server}
\format{\code{\link{R6Class}} object.}
\usage{
class_db_table_prototype
}
\description{
Class providing object with methods for communication with lightning-viz server
}
\section{Fields}{

\describe{
\item{\code{con}}{}
}}

\section{Methods}{


\describe{
  \item{???}{???}
}
}

\keyword{data}
