% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{DBpub}
\alias{DBprv}
\alias{DBpub}
\title{Database connector objects}
\format{R6 class generator objects}
\usage{
DBpub

DBprv
}
\description{
Implementation of two R6 classes and their generators for creating database connector objects.
Class "DBpub" implements all standard database interaction methods
(as in package \pkg{DBI}) as \emph{public} methods.
Class "DBprv" is identical, but with these methods implemented as \emph{private} methods.
}
\section{Methods}{

Implemented methods:
\describe{
\item{get_query(...)}{execute arbitrary SQL query, see \code{\link{dbGetQuery}}}
\item{list_tables(...)}{list tables in DB}
\item{list_fields(...)}{list fields of a table}
\item{send_query(...)}{send query}
\item{write_table(...)}{write a data frame to a DB table}
\item{read_table(...)}{read a DB table, return a data frame}
\item{has_tables(...)}{check if tables exist}
}
}
\examples{
z <- DBpub$new( DBI::dbConnect(RSQLite::SQLite(), dbname=":memory:") )
z$show()
z$list_tables()
z$write_table(name="mtcars", value=mtcars)
z$list_tables()
z$read_table("mtcars")
z$disconnect()
z$show()
}
\keyword{data}
\keyword{datasets}

