% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_db6.R
\docType{package}
\name{db6}
\alias{db6}
\alias{db6-package}
\title{db6}
\description{
db6
}
\section{Classes}{


\describe{
  \item{db}{A atabase class. (\link{class_db})}
  \item{db_con}{A wrapper around DBIConnection objects.  (\link{class_db_con})}
}
}

