% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_info.R
\name{db_get_info}
\alias{db_get_info}
\alias{db_get_info_SQLiteConnection}
\alias{db_get_info_AnsiConnection}
\alias{db_get_info_PostgreSQLConnection}
\title{db_get_info}
\usage{
db_get_info(con)

db_get_info_SQLiteConnection(con)

db_get_info_AnsiConnection(con)

db_get_info_PostgreSQLConnection(con)
}
\arguments{
\item{con}{}
}
\description{
db_get_info
}
