% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dbcon.R
\docType{class}
\name{Dbcon}
\alias{Dbcon}
\title{Class providing object with methods for communication with lightning-viz server}
\format{\code{\link{R6Class}} object.}
\usage{
Dbcon
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Class providing object with methods for communication with lightning-viz server
}
\section{Fields}{

\describe{
\item{\code{con}}{}
}}

\section{Methods}{


\describe{
  \item{\code{gallery(imgpathvector)}}{This method uploads images from vector of file paths \code{imgpathvector} to the server and creates a gallery of these images.}
  }
}

\keyword{data}
