% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R, R/new_db_con.R
\name{db_connect}
\alias{db_connect}
\alias{new_db_con}
\title{db_connect}
\usage{
db_connect(drv_fun, ...)

new_db_con(drv_fun, ...)
}
\arguments{
\item{drv_fun}{A function returning a database driver - e.g.: DBI::ANSI,
RSQLite::SQLite, ...}

\item{...}{paremeter passed through \code{class_db_con$new(...)} which passes
them through to DBI::dbConnect(drv_fun(), ...) - see \link[DBI]{dbConnect}.}
}
\description{
db_connect

new_db_con
}
