% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_db_table_prototype.R
\name{new_db_table_prototype}
\alias{new_db_table_prototype}
\title{new_db_table_prototype}
\usage{
new_db_table_prototype(db_con)
}
\arguments{
\item{db_con}{a object of class db_con -- see \link{new_db_con}, \link{class_db_con}}
}
\description{
new_db_table_prototype
}
