% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_extract_methods.R
\name{r6_extract_methods}
\alias{r6_extract_methods}
\title{r6_extract_methods}
\usage{
r6_extract_methods(r6)
}
\arguments{
\item{r6}{R6 object instance}
}
\description{
Extract public methods from R6 object instances
}
