% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{class}
\name{DB}
\alias{DB}
\alias{db}
\title{Database connector object}
\format{An R6 class generator object}
\usage{
DB

db(drv, ...)
}
\arguments{
\item{drv}{database driver object, see \code{\link{dbConnect}}}
}
\description{
Creating and using database connector objects.
}
\details{
Function \code{db} is used
}
\examples{
z <- db( RSQLite::SQLite(), dbname=":memory:")
z$show()
z$list_tables()
z$write_table(name="mtcars", value=mtcars)
z$list_tables()
z$read_table("mtcars")
z$disconnect()
z$show()
}
\keyword{data}

