% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_db_con.R
\name{new_db_con}
\alias{new_db_con}
\title{new_db_con}
\usage{
new_db_con(...)
}
\arguments{
\item{...}{paremeter passed through \code{class_db_con$new(...)} which passes
them through to DBI::dbConnect()}
}
\description{
new_db_con
}
